requirements.txt
dash==3.1.0
dash-bootstrap-components==2.0.3
plotly==6.2.0
pandas==2.3.0


langchain==0.3.26
langchain-community==0.3.26

opencv-python==4.11.0.86
Pillow==11.2.1

python-dotenv==1.1.1
tqdm==4.67.1
requests==2.32.4

llama-cpp-python==0.3.9
transformers==4.53.0
PyMuPDF==1.26.1
